+++ tools/SystemView/system/nordic/nrf.h
@@ -160,10 +160,10 @@ POSSIBILITY OF SUCH DAMAGE.
     #include "nrf52_to_nrf52833.h"
     #include "nrf51_to_nrf52.h"
 #elif defined (NRF52840_XXAA)
-    #include "nrf52840.h"
-    #include "nrf52840_bitfields.h"
-    #include "nrf51_to_nrf52840.h"
-    #include "nrf52_to_nrf52840.h"
+    #include "../nordic/nrf52840.h"
+    #include "../nordic/nrf52840_bitfields.h"
+    #include "../nordic/nrf51_to_nrf52840.h"
+    #include "../nordic/nrf52_to_nrf52840.h"
 
 #elif defined (NRF5340_XXAA)
     #if defined(NRF_APPLICATION)
@@ -208,7 +208,7 @@ POSSIBILITY OF SUCH DAMAGE.
     #error "Device must be defined. See nrf.h."
 #endif /* NRF51, NRF52805_XXAA, NRF52810_XXAA, NRF52811_XXAA, NRF52820_XXAA, NRF52832_XXAA, NRF52832_XXAB, NRF52833_XXAA, NRF52840_XXAA, NRF5340_XXAA_APPLICATION, NRF5340_XXAA_NETWORK, NRF9160_XXAA */
 
-#include "compiler_abstraction.h"
+#include "../nordic/compiler_abstraction.h"
 
 #endif /* NRF_H */
