+++ tools/SystemView/system/nordic/nrf52840.h
@@ -148,7 +148,7 @@ typedef enum {
 /** @} */ /* End of group Configuration_of_CMSIS */
 
 #include "core_cm4.h"                           /*!< ARM Cortex-M4 processor and core peripherals                              */
-#include "system_nrf52840.h"                    /*!< nrf52840 System                                                           */
+#include "../nordic/system_nrf52840.h"                    /*!< nrf52840 System                                                           */
 
 #ifndef __IM                                    /*!< Fallback for older CMSIS versions                                         */
   #define __IM   __I
