+++ tools/SystemView/system/nordic/nrf52840_bitfields.h
