+++ tools/SystemView/system/nordic/nrf52840_peripherals.h
