+++ tools/SystemView/system/nordic/nrf52_erratas.h
@@ -37,7 +37,8 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #include <stdint.h>
 #include <stdbool.h>
-#include "compiler_abstraction.h"
+
+#include "../nordic/compiler_abstraction.h"
 
 static bool nrf52_errata_1(void) __UNUSED;
 static bool nrf52_errata_2(void) __UNUSED;
