+++ tools/SystemView/system/nordic/system_nrf52.h
