+++ tools/SystemView/system/nrf52840.ld
@@ -261,7 +261,7 @@ SECTIONS
 
     /* Set stack top to end of RAM, and stack limit move down by
      * size of stack_dummy section */
-    __StackTop = ORIGIN(ram) + LENGTH(ram) - 0x5000;
+    __StackTop = ORIGIN(ram) + LENGTH(ram);
     __StackLimit = __StackTop - SIZEOF(.stack_dummy);
     __StackSize = SIZEOF(.stack_dummy);
     PROVIDE(__stack = __StackTop);
